Simplify OpenAPI auth extraction tests
